refactor: start transition to hexagonal architecture
