+++ transaction-service/src/main/java/com/financialmonitoring/transactionservice/config/mapper/MapperConfig.java
@@ -1,6 +1,6 @@
 package com.financialmonitoring.transactionservice.config.mapper;
 
-import com.financialmonitoring.transactionservice.config.mapper.custom.TransactionMapper;
+import com.financialmonitoring.transactionservice.config.mapper.converter.mapper.TransactionConverter;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -8,16 +8,16 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class MapperConfig {
 
-    private final TransactionMapper transactionMapper;
+    private final TransactionConverter transactionConverter;
 
-    public MapperConfig(TransactionMapper transactionMapper) {
-        this.transactionMapper = transactionMapper;
+    public MapperConfig(TransactionConverter transactionConverter) {
+        this.transactionConverter = transactionConverter;
     }
 
     @Bean
     public ModelMapper mapper() {
         ModelMapper mapper =  new ModelMapper();
-        mapper.addConverter(transactionMapper);
+        mapper.addConverter(transactionConverter);
         return mapper;
     }
 }
