+++ transaction-service/src/main/java/com/financialmonitoring/transactionservice/core/service/EventService.java
@@ -1,15 +1,15 @@
 package com.financialmonitoring.transactionservice.core.service;
 
-import com.financialmonitoring.transactionservice.core.model.Event;
-import com.financialmonitoring.transactionservice.core.repository.EventRepository;
+import com.financialmonitoring.transactionservice.infra.model.Event;
+import com.financialmonitoring.transactionservice.infra.repository.jpa.EventJpaRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 public class EventService {
-    private final EventRepository eventRepository;
+    private final EventJpaRepository eventJpaRepository;
 
-    public EventService(EventRepository eventRepository) {
-        this.eventRepository = eventRepository;
+    public EventService(EventJpaRepository eventJpaRepository) {
+        this.eventJpaRepository = eventJpaRepository;
     }
 
     public void notifyEnding(Event event) {
