+++ transaction-service/src/main/java/com/financialmonitoring/transactionservice/infra/repository/jpa/TransactionJpaRepository.java
@@ -1,0 +1,10 @@
+package com.financialmonitoring.transactionservice.infra.repository.jpa;
+
+import com.financialmonitoring.transactionservice.infra.model.Transaction;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TransactionJpaRepository extends MongoRepository<Transaction, String> {
+
+}
